% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump_dat.R
\name{clump_dat}
\alias{clump_dat}
\title{对Twosample包暴露格式数据进行去除连锁不平衡}
\usage{
clump_dat(
  dat,
  clump_kb = 10000,
  clump_r2 = 0.001,
  clump_p1 = 1,
  clump_p2 = 1,
  pop = "EUR",
  bfile = NULL,
  plink_bin = NULL
)
}
\arguments{
\item{dat}{Twosample包暴露格式数据。}

\item{clump_kb}{clump窗口大小,默认为10000。}

\item{clump_r2}{r2  clump阈值,默认为0.001。}

\item{clump_p1}{Clumping sig level for index SNPs, default is \code{1}.}

\item{clump_p2}{Clumping sig level for secondary SNPs, default is \code{1}.}

\item{pop}{人群种族,默认为"EUR"。}

\item{bfile}{如使用本地方式去除连锁不平衡，则需要提供1000人类基因组参考文件路径；如果在线方式，调用API则默认为NULL。}

\item{plink_bin}{如使用本地方式去除连锁不平衡，则需要提供plink软件的路径；如果在线方式，调用API则默认为NULL。}
}
\value{
data
}
\description{
对Twosample包暴露格式数据进行去除连锁不平衡
}
