% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_instrument.R
\name{extract_instrument}
\alias{extract_instrument}
\title{在线获取暴露数据}
\usage{
extract_instrument(
  outcomes,
  p1 = 5e-08,
  clump = TRUE,
  p2 = 5e-08,
  r2 = 0.001,
  kb = 10000,
  force_server = FALSE
)
}
\arguments{
\item{outcomes}{详细参考extract_instruments函数。}
}
\value{
data
}
\description{
在线获取暴露数据
}
